requirements.txt
Flask==2.0.1
google-cloud-aiplatform==1.70.0
vertexai-1.70.0
